Fix bug of upload photo and organize files
