+++ src/index.tsx
@@ -3,11 +3,11 @@ import ReactDOM from "react-dom/client"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 import "./index.css"
 import App from "./App"
-import Home from "./pages/landing-leaflet"
-import MyMap from "./pages/User/myMap-leaflet"
+import Home from "./pages/landing"
+import MyMap from "./pages/User/myMap"
 import MyMemories from "./pages/User/myMemories"
 import MyFriends from "./pages/User/myFriends"
-import FriendsMap from "./pages/User/friendMap-leaflet"
+import FriendsMap from "./pages/User/friendMap"
 import FriendMemories from "./pages/User/friendMemories"
 
 import reportWebVitals from "./reportWebVitals"
