+++ src/pages/Components/pinContent/detailMemory.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useContext, useRef, Dispatch } from "react"
 import { StreetViewService, GoogleMap, Marker } from "@react-google-maps/api"
 import styled from "styled-components"
 import parse from "html-react-parser"
-import { db, storage } from "../Utils/firebase"
+import { db, storage } from "../../Utils/firebase"
 import { doc, updateDoc, arrayUnion, arrayRemove } from "firebase/firestore"
 import { ref, deleteObject } from "firebase/storage"
 import { DocumentData } from "@firebase/firestore-types"
@@ -13,17 +13,17 @@ import {
   deleteMsg,
   PinContent,
   checkRealTimePhotos,
-} from "../User/functions/pins"
-import { AuthContext } from "../Context/authContext"
-import Editor from "../Components/editor"
-import Upload from "../User/components/uploadPhoto"
-import moreIcon from "../assets/buttons/moreIcon.png"
-import moreHoverIcon from "../assets/buttons/moreHover.png"
+} from "../../User/functions/pins"
+import { AuthContext } from "../../Context/authContext"
+import Editor from "../post/editor"
+import Upload from "../post/uploadPhoto"
 import SwiperPhotos from "./swiperPhoto"
-import whiteEditPencil from "../assets/buttons/edit.png"
-import blackEditPencil from "../assets/buttons/blackEdit.png"
-import calendar from "../assets/calendar.png"
-import location from "../assets/location.png"
+import moreIcon from "../../assets/buttons/moreIcon.png"
+import moreHoverIcon from "../../assets/buttons/moreHover.png"
+import whiteEditPencil from "../../assets/buttons/edit.png"
+import blackEditPencil from "../../assets/buttons/blackEdit.png"
+import calendar from "../../assets/calendar.png"
+import location from "../../assets/location.png"
 
 const Container = styled.div`
   position: absolute;
@@ -387,6 +387,7 @@ export default function DetailMemory(props: Props) {
   const [showUploadMore, setShowUploadMore] = useState(false)
   const [hasDiscard, setHasDiscard] = useState(false)
   const [selectedMsgId, setSelectedMsgId] = useState("")
+  const [canUpload, setCanUpload] = useState(true)
   const overlayRef = useRef<HTMLDivElement>(null)
 
   const updateTitle = async () => {
@@ -805,6 +806,7 @@ export default function DetailMemory(props: Props) {
                       typeof selectedMarker.location.placeId === "string" &&
                       (showEditor || showUploadMore) && (
                         <Upload
+                          canUpload={canUpload}
                           currentPin={{
                             id: selectedMarker.id,
                             userId: selectedMarker.userId,
