+++ src/pages/Components/pinContent/swiperPhoto.tsx
