+++ src/pages/Components/post/editor.tsx
