+++ src/pages/Components/profile.tsx
@@ -6,7 +6,7 @@ import imageCompression from "browser-image-compression"
 import { ref, uploadBytesResumable, getDownloadURL } from "firebase/storage"
 import { doc, updateDoc } from "firebase/firestore"
 import { AuthContext } from "../Context/authContext"
-import { Container } from "../User/components/styles"
+import { Container } from "./styles/mapStyles"
 import logoutIcon from "../assets/buttons/logoutIcon.png"
 import editPencil from "../assets/buttons/edit.png"
 import spinner from "../assets/dotsSpinner.svg"
