+++ src/pages/Components/sampleContent.tsx
@@ -12,7 +12,7 @@ import pkcAvatar from "../assets/samplePhotos/pkcAvatar.jpg"
 import cyuAvatar from "../assets/samplePhotos/cyuAvatar.jpg"
 import laysAvatar from "../assets/samplePhotos/laysAvatar.jpg"
 import moreIcon from "../assets/buttons/moreIcon.png"
-import SwiperPhotos from "./swiperPhoto"
+import SwiperPhotos from "./pinContent/swiperPhoto"
 import calendar from "../assets/calendar.png"
 import location from "../assets/location.png"
