+++ src/pages/Components/styles/widgetStyles.tsx
