+++ src/pages/User/functions/pins.tsx
@@ -11,7 +11,6 @@ import {
   onSnapshot,
   arrayUnion,
   arrayRemove,
-  orderBy,
 } from "firebase/firestore"
 import { DocumentData } from "@firebase/firestore-types"
 import { UserInfoType } from "../../Context/authContext"
