+++ src/pages/User/myMap.tsx
@@ -31,19 +31,27 @@ import {
   StepTitle,
   Input,
   BtnText,
-} from "../Components/styles/formStyle"
+} from "../Components/styles/formStyles"
+import {
+  BgOverlay,
+  ReminderArea,
+  ReminderText,
+  BtnWrapper,
+  BtnLight,
+  BtnBlue,
+} from "../Components/styles/popupLayoutStyles"
 import home from "../assets/markers/home1.png"
 import { StandaloneSearchBox } from "@react-google-maps/api"
 import { AuthContext } from "../Context/authContext"
-import Upload from "./components/uploadPhoto"
+import Upload from "../Components/post/uploadPhoto"
 import { db, storage } from "../Utils/firebase"
 import { doc, setDoc, updateDoc } from "firebase/firestore"
 import { ref, deleteObject } from "firebase/storage"
 import { getPins, PinContent } from "./functions/pins"
-import Editor from "../Components/editor"
+import Editor from "../Components/post/editor"
 import addPinIcon from "../assets/markers/addPin.png"
 import pins from "../assets/markers/pins.png"
-import DetailMemory from "../Components/detailMemory"
+import DetailMemory from "../Components/pinContent/detailMemory"
 import spinner from "../assets/dotsSpinner.svg"
 import xmark from "../assets/buttons/x-mark.png"
 
@@ -112,37 +120,6 @@ const BtnAddPin = styled.div`
   }
 `
 
-const BtnWrapper = styled.div`
-  display: flex;
-  flex: 1 1 auto;
-  width: 100%;
-  margin: 30px auto;
-
-  justify-content: space-between;
-  align-self: center;
-`
-const BtnBlue = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  width: 48%;
-  padding: 5px;
-  font-family: "Poppins";
-  font-size: 20px;
-  color: #ffffff;
-  background-color: ${(props) => props.theme.btnColor.bgBlue};
-  border-radius: 5px;
-  cursor: pointer;
-  @media screen and (max-width: 900px) and (min-width: 600px),
-    (max-height: 600px) {
-    font-size: 16px;
-  }
-`
-const BtnRed = styled(BtnBlue)`
-  background-color: ${(props) => props.theme.btnColor.bgRed};
-`
-
 const ArticleWrapper = styled.div`
   position: relative;
   display: flex;
@@ -172,36 +149,6 @@ export const PinInfoTitle = styled.div`
   font-weight: 700;
 `
 
-const BgOverlay = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: ${(props) => props.theme.color.bgDark};
-  opacity: 0.9;
-  z-index: 50;
-`
-const ReminderArea = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 50%;
-  padding: 20px;
-  color: ${(props) => props.theme.color.bgDark};
-  background-color: #fff;
-  border-radius: 5px;
-  z-index: 52;
-`
-const ReminderText = styled.div`
-  margin: 20px auto 0 auto;
-  text-align: center;
-  font-size: ${(props) => props.theme.title.lg};
-  @media screen and (max-width: 600px), (max-height: 600px) {
-    font-size: ${(props) => props.theme.title.md};
-  }
-`
 interface CenterType {
   center: LatLng | null
   setCenter: Dispatch<SetStateAction<LatLng | null>>
@@ -399,11 +346,12 @@ export default function MyMap() {
   const [showMemory, setShowMemory] = useState(false)
   const [showPostArea, setShowPostArea] = useState(false)
   const [showAlert, setShowAlert] = useState(false)
+  const [canUpload, setCanUpload] = useState(false)
   const overlayRef = useRef<HTMLDivElement>(null)
   const locationRef = useRef<HTMLInputElement>(null)
   const [refReady, setRefReady] = useState(false)
   const popupRef = useRef<any>(null)
-  console.log(locationRef?.current?.value)
+
   useEffect(() => {
     if (!selectedMarker) return
     if (refReady && popupRef !== undefined) {
@@ -458,6 +406,7 @@ export default function MyMap() {
           }
           setNewPin(newPinInfo)
           setHasAddPin(true)
+          setCanUpload(true)
         }
       }
     } else console.log("失敗啦")
@@ -497,8 +446,8 @@ export default function MyMap() {
               travelDate: travelDate,
               content: artiContent,
             },
-            postTimestamp: Date.now(),
-            postReadableTime: new Date(),
+            postTimestamp: artiInfo.postTimestamp,
+            postReadableTime: artiInfo.postReadableTime,
             albumURLs: urls,
             albumNames: filesName,
             location: {
@@ -518,8 +467,8 @@ export default function MyMap() {
           travelDate: travelDate,
           content: artiContent,
         },
-        postTimestamp: Date.now(),
-        postReadableTime: new Date(),
+        postTimestamp: artiInfo.postTimestamp,
+        postReadableTime: artiInfo.postReadableTime,
         albumURLs: urls,
         albumNames: filesName,
         location: {
@@ -540,6 +489,7 @@ export default function MyMap() {
       setTravelDate(getCurrentDate)
       setArtiContent("")
       setHasAddPin(false)
+      setCanUpload(false)
     } catch (error) {
       console.log(error)
     }
@@ -593,26 +543,25 @@ export default function MyMap() {
               <ReminderArea>
                 <ReminderText>
                   Are you sure <br />
-                  you want to discard all changes <br />
-                  and edit later?
+                  you want to discard this post?
                 </ReminderText>
 
                 <BtnWrapper>
-                  <BtnRed
+                  <BtnLight
+                    onClick={() => {
+                      setShowAlert(false)
+                    }}
+                  >
+                    cancel
+                  </BtnLight>
+                  <BtnBlue
                     onClick={() => {
                       cancelPost()
                       setShowAlert(false)
                       // setShowPostArea(false)
                     }}
                   >
-                    Yes
-                  </BtnRed>
-                  <BtnBlue
-                    onClick={() => {
-                      setShowAlert(false)
-                    }}
-                  >
-                    No, back to edit
+                    Discard
                   </BtnBlue>
                 </BtnWrapper>
               </ReminderArea>
@@ -689,7 +638,7 @@ export default function MyMap() {
                     setArtiContent={setArtiContent}
                   />
                   <Upload
-                    hasAddPin={hasAddPin}
+                    canUpload={canUpload}
                     currentPin={newPin}
                     setFilesName={setFilesName}
                     hasUpload={hasUpload}
