+++ src/pages/Widgets/weather.tsx
@@ -18,13 +18,13 @@ import {
 } from "chart.js"
 import { Line } from "react-chartjs-2"
 import { AuthContext } from "../Context/authContext"
-import { Input } from "../Components/styles/formStyle"
+import { Input } from "../Components/styles/formStyles"
 import {
   GridContainer,
   GridItemWrapper,
   Xmark,
   FormTitle,
-} from "../Components/styles/gridStyles"
+} from "../Components/styles/widgetStyles"
 import spinner from "../assets/dotsSpinner.svg"
 
 const PlaceInput = styled(Input)`
